# scaffold/hr-service/openapi.yaml
openapi: 3.0.3
info:
  title: PMP Human Resources Service
  version: "1.0"
  description: >
    Implements FR-HR-01..07 (docs Section 4.6). FastAPI must generate a schema
    that matches this contract — treat mismatches as a bug, not a spec update,
    unless the SRS itself is being revised.

    All endpoints require a bearer access token (RS256 JWT) issued by
    iam-service; its signature is verified against iam-service's JWKS and the
    required `<resource>.<action>` permission must be present in the token's
    `permissions` claim (CLAUDE.md rule 4). Missing/invalid token -> 401;
    valid token without the permission -> 403.

    discipline_records are additionally restricted: only hr.discipline.read /
    hr.discipline.write unlock those routes (docs Section 2.3 — HR/command
    only). FR-HR-08 (roster/headcount reports) is not implemented — it reads
    to derive reports rather than defining new domain state, and is deferred.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /officers/{officer_id}/promotions:
    post:
      summary: Record a promotion (FR-HR-04)
      description: >
        Requires `hr.promotion.write`. No approval workflow — `previous_rank`
        is the officer's current rank at the time of recording; the officer's
        rank is updated immediately. Publishes a `PromotionRecorded` domain
        event.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: officer_id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PromotionCreate" }
      responses:
        "201": { description: Recorded, content: { application/json: { schema: { $ref: "#/components/schemas/Promotion" } } } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.promotion.write }
        "404": { description: No officer with that id }
    get:
      summary: Promotion history for one officer (FR-HR-04)
      description: Requires `hr.promotion.read`.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: officer_id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Promotions, newest first
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Promotion" } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.promotion.read }
        "404": { description: No officer with that id }
...
  /officers/{officer_id}/performance-reviews:
    post:
      summary: Record a performance review (FR-HR-07)
      description: >
        Requires `hr.performance.write`. Publishes a
        `PerformanceReviewRecorded` domain event.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: officer_id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PerformanceReviewCreate" }
      responses:
        "201": { description: Recorded, content: { application/json: { schema: { $ref: "#/components/schemas/PerformanceReview" } } } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.performance.write }
        "404": { description: No officer with that id, or reviewer_id does not exist }
    get:
      summary: Performance-review history for one officer (FR-HR-07)
      description: Requires `hr.performance.read`.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: officer_id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Performance reviews, newest first
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/PerformanceReview" } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.performance.read }
        "404": { description: No officer with that id }
  /performance-reviews/{performance_review_id}:
    get:
      summary: Retrieve one performance review (FR-HR-07)
      description: Requires `hr.performance.read`.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: performance_review_id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/PerformanceReview" } } } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.performance.read }
        "404": { description: No performance review with that id }
    patch:
      summary: Update a performance review (FR-HR-07)
      description: >
        Requires `hr.performance.write`. Publishes a
        `PerformanceReviewUpdated` domain event.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: performance_review_id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PerformanceReviewUpdate" }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/PerformanceReview" } } } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.performance.write }
        "404": { description: No performance review with that id }
    delete:
      summary: Delete a performance review (FR-HR-07)
      description: >
        Requires `hr.performance.write`. Publishes a
        `PerformanceReviewDeleted` domain event.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: performance_review_id, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Deleted }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.performance.write }
        "404": { description: No performance review with that id }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 access token issued by iam-service (/api/v1/auth/mfa/verify).
  schemas:
...
    PromotionCreate:
      type: object
      required: [new_rank]
      properties:
        new_rank: { type: string }
    Promotion:
      allOf:
        - $ref: "#/components/schemas/PromotionCreate"
        - type: object
          properties:
            id: { type: string, format: uuid }
            officer_id: { type: string, format: uuid }
            previous_rank: { type: string }
            created_at: { type: string, format: date-time }
...
    PerformanceReviewCreate:
      type: object
      required: [reviewer_id, period, score]
      properties:
        reviewer_id: { type: string, format: uuid }
        period: { type: string }
        score: { type: number, format: float, minimum: 0, maximum: 99.99 }
    PerformanceReview:
      allOf:
        - $ref: "#/components/schemas/PerformanceReviewCreate"
        - type: object
          properties:
            id: { type: string, format: uuid }
            officer_id: { type: string, format: uuid }
            created_at: { type: string, format: date-time }
    PerformanceReviewUpdate:
      type: object
      properties:
        period: { type: string, nullable: true }
        score: { type: number, format: float, minimum: 0, maximum: 99.99, nullable: true }